+++ database/migrations/2014_07_09_154610_add_deleted_to_spice_harvester_records.php
@@ -12,17 +12,17 @@ class AddDeletedToSpiceHarvesterRecords extends Migration {
 	 */
 	public function up()
 	{
-		Schema::table('spice_harvester_records', function($table)
-		{
-		    // $table->boolean('is_deleted')->default(false);
-		    $table->softDeletes();
+		Schema::table('spice_harvester_records', function ($table) {
+			// $table->boolean('is_deleted')->default(false);
+			$table->softDeletes();
 
-		    // $table->leftJoin('items', 'spice_harvester_records.item_id', '=', 'items.id')->whereRaw('items.id IS NULL')->update(array('deleted_at' => date("Y-m-d H:i:s")));
+			// $table->leftJoin('items', 'spice_harvester_records.item_id', '=', 'items.id')->whereRaw('items.id IS NULL')->update(array('deleted_at' => date("Y-m-d H:i:s")));
 		});
 
-		//tie zaznamy, pre ktore uz neexistuju diela - nastavit ako zmazane
-	    DB::unprepared( 'UPDATE spice_harvester_records r LEFT JOIN items i ON i.id = r.item_id SET r.deleted_at = NOW() WHERE i.id IS NULL;' );
-
+		if (!DB::connection() instanceof \Illuminate\Database\SQLiteConnection) {
+			//tie zaznamy, pre ktore uz neexistuju diela - nastavit ako zmazane
+			DB::unprepared('UPDATE spice_harvester_records r LEFT JOIN items i ON i.id = r.item_id SET r.deleted_at = NOW() WHERE i.id IS NULL;');
+		}
 	}
 
 	/**