+++ database/migrations/2014_07_21_110004_add_idenitifier_to_item.php
@@ -14,7 +14,7 @@ class AddIdenitifierToItem extends Migration {
 	{
 		Schema::table('items', function($table)
 		{
-			$table->string('identifier');
+			$table->string('identifier')->default('');
 		});
 	}