+++ database/migrations/2017_11_16_172658_add_class_name_to_imports_table.php
@@ -13,7 +13,7 @@ class AddClassNameToImportsTable extends Migration
     public function up()
     {
         Schema::table('imports', function (Blueprint $table) {
-            $table->string('class_name');
+            $table->string('class_name')->default('');
         });
     }